refactor: move firehose role helper
